+++ temporian/implementation/numpy/operators/__init__.py
@@ -64,6 +64,7 @@ from temporian.implementation.numpy.operators import map
 from temporian.implementation.numpy.operators import select_index_values
 from temporian.implementation.numpy.operators import since_last
 from temporian.implementation.numpy.operators import tick
+from temporian.implementation.numpy.operators import tick_calendar
 from temporian.implementation.numpy.operators import timestamps
 from temporian.implementation.numpy.operators import unique_timestamps
 from temporian.implementation.numpy.operators import until_next