+++ tools/start_compile_docker.sh
@@ -4,4 +4,4 @@
 
 DOCKER=gcr.io/tfx-oss-public/manylinux2014-bazel:bazel-5.3.0
 TEMPORIAN_DIR=temporian
-sudo docker run -it -v ${PWD}/..:/working_dir -w /working_dir/${TEMPORIAN_DIR} ${DOCKER} $@
+sudo docker run -it --rm -v ${PWD}/..:/working_dir -w /working_dir/${TEMPORIAN_DIR} ${DOCKER} $@